# pkg/apiclient/api/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: Daytona Server API
  title: Daytona Server API
  version: 0.1.0
servers:
- url: http://localhost:3986/
security:
- Bearer: []
paths:
...
  /project-config:
    get:
      description: List project configs
      operationId: ListProjectConfigs
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectConfig'
                type: array
          description: OK
      summary: List project configs
      tags:
      - project-config
    put:
      description: Set project config data
      operationId: SetProjectConfig
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectConfigDTO'
        description: Project config
        required: true
      responses:
        "201":
          content: {}
          description: Created
      summary: Set project config data
      tags:
      - project-config
      x-codegen-request-body-name: projectConfig
  /project-config/{configName}:
    delete:
      description: Delete project config data
      operationId: DeleteProjectConfig
      parameters:
      - description: Config name
        in: path
        name: configName
        required: true
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: No Content
      summary: Delete project config data
      tags:
      - project-config
    get:
      description: Get project config data
      operationId: GetProjectConfig
      parameters:
      - description: Config name
        in: path
        name: configName
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProjectConfig'
          description: OK
      summary: Get project config data
      tags:
      - project-config
...
components:
  schemas:
...
    CreateProjectConfigDTO:
      example:
        image: image
        build:
          devcontainer:
            devContainerFilePath: devContainerFilePath
        envVars:
          key: envVars
        name: name
        source:
          repository:
            owner: owner
            path: path
            name: name
            id: id
            source: source
            prNumber: 0
            branch: branch
            sha: sha
            url: url
        user: user
      properties:
        build:
          $ref: '#/components/schemas/ProjectBuild'
        envVars:
          additionalProperties:
            type: string
          type: object
        image:
          type: string
        name:
          type: string
        source:
          $ref: '#/components/schemas/CreateProjectConfigSourceDTO'
        user:
          type: string
      type: object
    CreateProjectConfigSourceDTO:
      example:
        repository:
          owner: owner
          path: path
          name: name
          id: id
          source: source
          prNumber: 0
          branch: branch
          sha: sha
          url: url
      properties:
        repository:
          $ref: '#/components/schemas/GitRepository'
      type: object
...
    GitRepository:
      example:
        owner: owner
        path: path
        name: name
        id: id
        source: source
        prNumber: 0
        branch: branch
        sha: sha
        url: url
      properties:
        branch:
          type: string
        id:
          type: string
        name:
          type: string
        owner:
          type: string
        path:
          type: string
        prNumber:
          type: integer
        sha:
          type: string
        source:
          type: string
        url:
          type: string
      type: object
...
    ProjectBuild:
      example:
        devcontainer:
          devContainerFilePath: devContainerFilePath
      properties:
        devcontainer:
          $ref: '#/components/schemas/ProjectBuildDevcontainer'
      type: object
    ProjectBuildDevcontainer:
      example:
        devContainerFilePath: devContainerFilePath
      properties:
        devContainerFilePath:
          type: string
      type: object
    ProjectConfig:
      example:
        image: image
        build:
          devcontainer:
            devContainerFilePath: devContainerFilePath
        name: name
        repository:
          owner: owner
          path: path
          name: name
          id: id
          source: source
          prNumber: 0
          branch: branch
          sha: sha
          url: url
        user: user
      properties:
        build:
          $ref: '#/components/schemas/ProjectBuild'
        image:
          type: string
        name:
          type: string
        repository:
          $ref: '#/components/schemas/GitRepository'
        user:
          type: string
      type: object
...
  securitySchemes:
    Bearer:
      description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
      in: header
      name: Authorization
      type: apiKey
...